global package buying api added

+++ database/migrations/2021_12_19_140904_create_therapist_locations_table.php
@@ -16,7 +16,7 @@ class CreateTherapistLocationsTable extends Migration
         Schema::create('therapist_locations', function (Blueprint $table) {
             $table->id();
             $table->string('address');
-            $table->integer('price');
+            $table->double('price');
             $table->integer('share');
             $table->integer('discount');
             $table->boolean('is_active')->default(true);
