+++ database/migrations/2021_12_20_153228_create_care_giver_services_table.php
@@ -17,7 +17,7 @@ class CreateCareGiverServicesTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('slug')->unique();
-            $table->integer('price');
+            $table->double('price');
             $table->integer('discount');
             $table->integer('share');
             $table->boolean('is_active')->default(true);
